+++ src/domain/SlackReminder.test.ts
@@ -20,12 +20,26 @@ describe('SlackReminder', () => {
         d0: new Day(new Date('2024-01-01')),
         size: 7,
         holidays: [],
+        users: [new Email('a@example.com')],
       }
     ).getMessage()
     deepEqual(message, {
       text: `See the 7-day vacation overview for \`test\`
 ▼ Today
 🌴🗓🗓🗓🗓🌞🌞 —a@example.com
+`,
+    })
+  })
+  test('Slack message displays "empty case" in case there are no users', () => {
+    const message = new SlackReminder('test', [], new SlackSubscriber('john'), {
+      d0: new Day(new Date('2024-01-01')),
+      size: 7,
+      holidays: [],
+      users: [],
+    }).getMessage()
+    deepEqual(message, {
+      text: `See the 7-day vacation overview for \`test\`
+_Tracker has no subjects_
 `,
     })
   })